+++ cmd/package-image/buildah.go
@@ -19,16 +19,31 @@ const (
 
 // buildahBuild builds a local image using the Dockerfile and context directory
 // given in opts, tagging the resulting image with given tag.
-func buildahBuild(opts options, tag string, outWriter, errWriter io.Writer) error {
-	args, err := buildahBuildArgs(opts, tag)
+func (p *packageImage) buildahBuild(outWriter, errWriter io.Writer) error {
+	args, err := p.buildahBuildArgs(p.image.Ref)
 	if err != nil {
 		return fmt.Errorf("assemble build args: %w", err)
 	}
 	return command.Run(buildahBin, args, []string{}, outWriter, errWriter)
 }
 
+// buildahPush pushes a local image to a OCI formatted directory for trivy image scans.
+func (p *packageImage) buildahPushTar(outWriter, errWriter io.Writer) error {
+	args := []string{
+		fmt.Sprintf("--storage-driver=%s", p.opts.storageDriver),
+		"push",
+		fmt.Sprintf("--digestfile=%s", filepath.Join(p.opts.checkoutDir, "image-digest")),
+	}
+	if p.opts.debug {
+		args = append(args, "--log-level=debug")
+	}
+	args = append(args, p.image.Ref, fmt.Sprintf("oci:%s", filepath.Join(p.opts.checkoutDir, p.image.Name)))
+	return command.Run(buildahBin, args, []string{}, outWriter, errWriter)
+}
+
 // buildahPush pushes a local image to the given imageRef.
-func buildahPush(opts options, workingDir, imageRef string, outWriter, errWriter io.Writer) error {
+func (p *packageImage) buildahPush(outWriter, errWriter io.Writer) error {
+	opts := p.opts
 	extraArgs, err := shlex.Split(opts.buildahPushExtraArgs)
 	if err != nil {
 		log.Printf("could not parse extra args (%s): %s", opts.buildahPushExtraArgs, err)
@@ -38,22 +53,23 @@ func buildahPush(opts options, workingDir, imageRef string, outWriter, errWriter
 		"push",
 		fmt.Sprintf("--tls-verify=%v", opts.tlsVerify),
 		fmt.Sprintf("--cert-dir=%s", opts.certDir),
-		fmt.Sprintf("--digestfile=%s", filepath.Join(workingDir, "image-digest")),
+		fmt.Sprintf("--digestfile=%s", filepath.Join(opts.checkoutDir, "image-digest")),
 	}
 	args = append(args, extraArgs...)
 	if opts.debug {
 		args = append(args, "--log-level=debug")
 	}
-	args = append(args, imageRef, fmt.Sprintf("docker://%s", imageRef))
+	args = append(args, p.image.Ref, fmt.Sprintf("docker://%s", p.image.Ref))
 	return command.Run(buildahBin, args, []string{}, outWriter, errWriter)
 }
 
 // buildahBuildArgs assembles the args to be passed to buildah based on
 // given options and tag.
-func buildahBuildArgs(opts options, tag string) ([]string, error) {
+func (p *packageImage) buildahBuildArgs(tag string) ([]string, error) {
 	if tag == "" {
 		return nil, errors.New("tag must not be empty")
 	}
+	opts := p.opts
 	extraArgs, err := shlex.Split(opts.buildahBuildExtraArgs)
 	if err != nil {
 		return nil, fmt.Errorf("parse extra args (%s): %w", opts.buildahBuildExtraArgs, err)
@@ -70,7 +86,7 @@ func buildahBuildArgs(opts options, tag string) ([]string, error) {
 		fmt.Sprintf("--tag=%s", tag),
 	}
 	args = append(args, extraArgs...)
-	nexusArgs, err := nexusBuildArgs(opts)
+	nexusArgs, err := p.nexusBuildArgs()
 	if err != nil {
 		return nil, fmt.Errorf("add nexus build args: %w", err)
 	}
@@ -85,8 +101,9 @@ func buildahBuildArgs(opts options, tag string) ([]string, error) {
 // nexusBuildArgs computes --build-arg parameters so that the Dockerfile
 // can access nexus as determined by the options nexus related
 // parameters.
-func nexusBuildArgs(opts options) ([]string, error) {
+func (p *packageImage) nexusBuildArgs() ([]string, error) {
 	args := []string{}
+	opts := p.opts
 	if strings.TrimSpace(opts.nexusURL) != "" {
 		nexusUrl, err := url.Parse(opts.nexusURL)
 		if err != nil {