+++ cmd/package-image/main.go
@@ -3,14 +3,11 @@ package main
 import (
 	"flag"
 	"fmt"
-	"log"
 	"os"
 	"path/filepath"
 	"strings"
 	"time"
 
-	"github.com/opendevstack/pipeline/internal/command"
-	"github.com/opendevstack/pipeline/internal/directory"
 	"github.com/opendevstack/pipeline/pkg/artifact"
 	"github.com/opendevstack/pipeline/pkg/bitbucket"
 	"github.com/opendevstack/pipeline/pkg/logging"
@@ -22,6 +19,7 @@ const (
 )
 
 type options struct {
+	checkoutDir           string
 	bitbucketAccessToken  string
 	bitbucketURL          string
 	aquaUsername          string
@@ -42,11 +40,20 @@ type options struct {
 	nexusPassword         string
 	buildahBuildExtraArgs string
 	buildahPushExtraArgs  string
+	trivySBOMExtraArgs    string
 	aquasecGate           bool
 	debug                 bool
 }
 
+type packageImage struct {
+	logger logging.LeveledLoggerInterface
+	opts   options
+	ctxt   *pipelinectxt.ODSContext
+	image  artifact.Image
+}
+
 var defaultOptions = options{
+	checkoutDir:           ".",
 	bitbucketAccessToken:  os.Getenv("BITBUCKET_ACCESS_TOKEN"),
 	bitbucketURL:          os.Getenv("BITBUCKET_URL"),
 	aquaUsername:          os.Getenv("AQUA_USERNAME"),
@@ -55,7 +62,7 @@ var defaultOptions = options{
 	aquaRegistry:          os.Getenv("AQUA_REGISTRY"),
 	imageStream:           "",
 	registry:              "image-registry.openshift-image-registry.svc:5000",
-	certDir:               "/etc/containers/certs.d",
+	certDir:               defaultCertDir(),
 	imageNamespace:        "",
 	tlsVerify:             true,
 	storageDriver:         "vfs",
@@ -67,12 +74,14 @@ var defaultOptions = options{
 	nexusPassword:         os.Getenv("NEXUS_PASSWORD"),
 	buildahBuildExtraArgs: "",
 	buildahPushExtraArgs:  "",
+	trivySBOMExtraArgs:    "",
 	aquasecGate:           false,
 	debug:                 (os.Getenv("DEBUG") == "true"),
 }
 
 func main() {
 	opts := options{}
+	flag.StringVar(&opts.checkoutDir, "checkout-dir", defaultOptions.checkoutDir, "Checkout dir")
 	flag.StringVar(&opts.bitbucketAccessToken, "bitbucket-access-token", defaultOptions.bitbucketAccessToken, "bitbucket-access-token")
 	flag.StringVar(&opts.bitbucketURL, "bitbucket-url", defaultOptions.bitbucketURL, "bitbucket-url")
 	flag.StringVar(&opts.aquaUsername, "aqua-username", defaultOptions.aquaUsername, "aqua-username")
@@ -93,126 +102,39 @@ func main() {
 	flag.StringVar(&opts.nexusPassword, "nexus-password", defaultOptions.nexusPassword, "Nexus password")
 	flag.StringVar(&opts.buildahBuildExtraArgs, "buildah-build-extra-args", defaultOptions.buildahBuildExtraArgs, "extra parameters passed for the build command when building images")
 	flag.StringVar(&opts.buildahPushExtraArgs, "buildah-push-extra-args", defaultOptions.buildahPushExtraArgs, "extra parameters passed for the push command when pushing images")
+	flag.StringVar(&opts.trivySBOMExtraArgs, "trivy-sbom-extra-args", defaultOptions.trivySBOMExtraArgs, "extra parameters passed for the trivy command to generate an SBOM")
 	flag.BoolVar(&opts.aquasecGate, "aqua-gate", defaultOptions.aquasecGate, "whether the Aqua security scan needs to pass for the task to succeed")
 	flag.BoolVar(&opts.debug, "debug", defaultOptions.debug, "debug mode")
 	flag.Parse()
 
-	workingDir := "."
-	ctxt := &pipelinectxt.ODSContext{}
-	err := ctxt.ReadCache(workingDir)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	if _, err := os.Stat(kubernetesServiceaccountDir); err == nil {
-		opts.certDir = kubernetesServiceaccountDir
-	}
+	var logger logging.LeveledLoggerInterface
 	if opts.debug {
-		err := directory.ListFiles(opts.certDir, os.Stdout)
-		if err != nil {
-			log.Fatal(err)
-		}
-	}
-
-	// TLS verification of the KinD registry is not possible at the moment as
-	// requests error out with "server gave HTTP response to HTTPS client".
-	if strings.HasPrefix(opts.registry, "kind-registry.kind") {
-		opts.tlsVerify = false
-	}
-
-	imageNamespace := opts.imageNamespace
-	if len(imageNamespace) == 0 {
-		imageNamespace = ctxt.Namespace
-	}
-	imageStream := opts.imageStream
-	if len(imageStream) == 0 {
-		imageStream = ctxt.Component
-	}
-	imageName := fmt.Sprintf("%s:%s", imageStream, ctxt.GitCommitSHA)
-	imageRef := fmt.Sprintf(
-		"%s/%s/%s",
-		opts.registry, imageNamespace, imageName,
-	)
-
-	fmt.Printf("Checking if image %s exists already ...\n", imageName)
-	imageDigest, err := getImageDigestFromRegistry(imageRef, opts)
-	if err == nil {
-		fmt.Println("Image exists already.")
+		logger = &logging.LeveledLogger{Level: logging.LevelDebug}
 	} else {
-
-		fmt.Printf("Building image %s ...\n", imageName)
-		err := buildahBuild(opts, imageRef, os.Stdout, os.Stderr)
-		if err != nil {
-			log.Fatal("buildah bud: ", err)
-		}
-
-		fmt.Printf("Pushing image %s ...\n", imageRef)
-		err = buildahPush(opts, workingDir, imageRef, os.Stdout, os.Stderr)
-		if err != nil {
-			log.Fatal("buildah push: ", err)
-		}
-
-		d, err := getImageDigestFromFile(workingDir)
-		if err != nil {
-			log.Fatal(err)
-		}
-		imageDigest = d
-
-		if aquasecInstalled() {
-			fmt.Println("Scanning image with Aqua scanner ...")
-			aquaImage := fmt.Sprintf("%s/%s", imageNamespace, imageName)
-			htmlReportFile := filepath.Join(workingDir, "report.html")
-			jsonReportFile := filepath.Join(workingDir, "report.json")
-			scanArgs := aquaAssembleScanArgs(opts, aquaImage, htmlReportFile, jsonReportFile)
-			scanSuccessful, err := aquaScan(aquasecBin, scanArgs, os.Stdout, os.Stderr)
-			if err != nil {
-				log.Fatal("aqua scan: ", err)
-			}
-
-			if !scanSuccessful && opts.aquasecGate {
-				log.Fatalln("Stopping build as successful Aqua scan is required")
-			}
-
-			asu, err := aquaScanURL(opts, aquaImage)
-			if err != nil {
-				log.Fatal("aqua scan URL:", err)
-			}
-			fmt.Printf("Aqua vulnerability report is at %s ...\n", asu)
-
-			err = copyAquaReportsToArtifacts(htmlReportFile, jsonReportFile)
-			if err != nil {
-				log.Fatal(err)
-			}
-
-			fmt.Println("Creating Bitbucket code insight report ...")
-			err = createBitbucketInsightReport(opts, asu, scanSuccessful, ctxt)
-			if err != nil {
-				log.Fatal(err)
-			}
-		} else {
-			fmt.Println("Aqua is not configured, image will not be scanned for vulnerabilities.")
-		}
+		logger = &logging.LeveledLogger{Level: logging.LevelInfo}
 	}
 
-	err = writeImageDigestToResults(imageDigest)
+	err := (&packageImage{logger: logger, opts: opts}).runSteps(
+		setupContext(),
+		setImageName(),
+		skipIfImageArtifactExists(),
+		buildImageAndGenerateTar(),
+		generateSBOM(),
+		pushImage(),
+		scanImageWithAqua(),
+		storeArtifact(),
+	)
 	if err != nil {
-		log.Fatal(err)
+		logger.Errorf(err.Error())
+		os.Exit(1)
 	}
+}
 
-	fmt.Println("Writing image artifact ...")
-	ia := artifact.Image{
-		Image:      imageRef,
-		Registry:   opts.registry,
-		Repository: imageNamespace,
-		Name:       imageStream,
-		Tag:        ctxt.GitCommitSHA,
-		Digest:     imageDigest,
+func defaultCertDir() string {
+	if _, err := os.Stat(kubernetesServiceaccountDir); err == nil {
+		return kubernetesServiceaccountDir
 	}
-	imageArtifactFilename := fmt.Sprintf("%s.json", imageStream)
-	err = pipelinectxt.WriteJsonArtifact(ia, pipelinectxt.ImageDigestsPath, imageArtifactFilename)
-	if err != nil {
-		log.Fatal(err)
-	}
+	return "/etc/containers/certs.d"
 }
 
 // copyAquaReportsToArtifacts copies the Aqua scan reports to the artifacts directory.
@@ -278,23 +200,6 @@ func createBitbucketInsightReport(opts options, aquaScanUrl string, success bool
 	return err
 }
 
-// getImageDigestFromRegistry returns a SHA256 image digest if the specified
-// imageRef exists. Example return value:
-// "sha256:3b6de1c737065e9973ddb7cc60b769b866b7649ff6f2de3816934dda832de294"
-func getImageDigestFromRegistry(imageRef string, opts options) (string, error) {
-	args := []string{
-		"inspect",
-		fmt.Sprintf("--format=%s", "{{.Digest}}"),
-		fmt.Sprintf("--tls-verify=%v", opts.tlsVerify),
-		fmt.Sprintf("--cert-dir=%s", opts.certDir),
-	}
-	if opts.debug {
-		args = append(args, "--debug")
-	}
-	stdout, _, err := command.RunBuffered("skopeo", append(args, "docker://"+imageRef))
-	return string(stdout), err
-}
-
 // getImageDigestFromFile reads the image digest from the file written to by buildah.
 func getImageDigestFromFile(workingDir string) (string, error) {
 	content, err := os.ReadFile(filepath.Join(workingDir, "image-digest"))
@@ -311,4 +216,12 @@ func writeImageDigestToResults(imageDigest string) error {
 		return err
 	}
 	return os.WriteFile("/tekton/results/image-digest", []byte(imageDigest), 0644)
+}
+
+// imageArtifactExists checks if image artifact JSON file exists in its artifacts path
+func imageArtifactExists(p *packageImage) error {
+	imageArtifactsDir := filepath.Join(p.opts.checkoutDir, pipelinectxt.ImageDigestsPath)
+	imageArtifactFilename := fmt.Sprintf("%s.json", p.ctxt.Component)
+	_, err := os.Stat(filepath.Join(imageArtifactsDir, imageArtifactFilename))
+	return err
 }